requirements.txt
slack-bolt==1.4.4
google-cloud-storage==1.37.1
gunicorn==20.1.0
aiohttp==3.7.4.post0
